+++ src/library/DIPS.Mobile.UI/Components/Pickers/TimePicker/TimePicker.Properties.cs
@@ -32,4 +32,32 @@ public partial class TimePicker
         get => (ICommand)GetValue(SelectedTimeCommandProperty);
         set => SetValue(SelectedTimeCommandProperty, value);
     }
+    
+    public static readonly BindableProperty MinimumTimeProperty = BindableProperty.Create(
+        nameof(MinimumTime),
+        typeof(TimeSpan?),
+        typeof(TimePicker));
+
+    /// <summary>
+    /// Sets the minimum time that people can choose.
+    /// </summary>
+    public TimeSpan? MinimumTime
+    {
+        get => (TimeSpan?)GetValue(MinimumTimeProperty);
+        set => SetValue(MinimumTimeProperty, value);
+    }
+    
+    public static readonly BindableProperty MaximumTimeProperty = BindableProperty.Create(
+        nameof(MaximumTime),
+        typeof(TimeSpan?),
+        typeof(TimePicker));
+
+    /// <summary>
+    /// Sets the maximum time that people can choose.
+    /// </summary>
+    public TimeSpan? MaximumTime
+    {
+        get => (TimeSpan?)GetValue(MaximumTimeProperty);
+        set => SetValue(MaximumTimeProperty, value);
+    }
 }